+++ src/Serilog.Sinks.File.Encrypt/Readers/v1/SessionReaderV1.cs
@@ -1,4 +1,3 @@
-using System.Security.Cryptography;
 using Serilog.Sinks.File.Encrypt.Interfaces;
 using Serilog.Sinks.File.Encrypt.Models;
 
@@ -21,28 +20,29 @@ internal class SessionReaderV1 : ISessionReader
     /// <inheritdoc />
     public async Task<DecryptionContext> ReadSessionAsync(
         Stream input,
-        Dictionary<string, RSA> keyMap,
+        IKeyProvider keyProvider,
         CancellationToken cancellationToken
     )
     {
+        ArgumentNullException.ThrowIfNull(input);
+        ArgumentNullException.ThrowIfNull(keyProvider);
         // Read the header from the input stream
         Memory<byte> keyId = new byte[HeaderMetadataV1.KeyIdLength];
         // lookup the RSA key based on the keyId in the header
         await input.ReadExactlyAsync(keyId, cancellationToken);
         string keyIdStr = System.Text.Encoding.UTF8.GetString(keyId.Span).TrimEnd('\0');
-        if (!keyMap.TryGetValue(keyIdStr, out RSA? rsa))
-        {
-            throw new InvalidOperationException(
-                $"No RSA private key found for KeyId: '{keyIdStr}'."
-            );
-        }
 
-        int headerSize = rsa.KeySize / 8;
+        int headerSize = await keyProvider.GetKeySizeAsync(keyIdStr, cancellationToken) / 8;
         Memory<byte> header = new byte[headerSize];
         await input.ReadExactlyAsync(header, cancellationToken);
 
         // Decrypt the header to get the session key and nonce
-        (byte[] aesKey, byte[] nonce) = _headerReader.Decrypt(rsa, header.Span);
+        (byte[] aesKey, byte[] nonce) = await _headerReader.Decrypt(
+            keyProvider,
+            keyIdStr,
+            header,
+            cancellationToken
+        );
 
         return new DecryptionContext(nonce, aesKey);
     }